Re-generate API types
- Enums seems to have been removed from the Swagger spec, however they were kept and are now in `src/types/enums.ts`
