+++ src/generated.ts
@@ -13,10 +13,11 @@ export type AcceptEuaData = EndUserAgreement;
 
 export type AcceptEuaError = ErrorResponse;
 
-/** The representation of a bank account. */
+/** AccountSerializer. */
 export interface Account {
+  /** The Account BBAN */
+  bban?: string;
   /**
-   * Created date/time
    * The date & time at which the account object was created.
    * @format date-time
    */
@@ -24,43 +25,35 @@ export interface Account {
   /** The Account IBAN */
   iban?: string;
   /**
-   * Account ID
    * The ID of this Account, used to refer to this account in other API calls.
    * @format uuid
    */
   id?: string;
-  /**
-   * ASPSP ID
-   * The ASPSP associated with this account.
-   */
+  /** The ASPSP associated with this account. */
   institution_id?: string;
   /**
-   * Last accessed date/time
    * The date & time at which the account object was last accessed.
    * @format date-time
    */
-  last_accessed?: string | null;
-  /**
-   * Account Owner Name
-   * The name of the account owner.
-   */
+  last_accessed?: string;
+  /** The name of the account owner. */
   owner_name?: string;
-  /**
-   * Account Status
-   * The processing status of this account.
-   */
-  status?: AccountStatusEnum;
+  /** The processing status of this account. */
+  status?: string;
 }
 
-export interface AccountBalanceSchema {
+/** AccountBalanceSerializer. */
+export interface AccountBalance {
   balances?: BalanceSchema[];
 }
 
-export interface AccountDetailSchema {
+/** AccountDetailSerializer. */
+export interface AccountDetail {
   /** account */
   account: DetailSchema;
 }
 
+/** AccountSchema. */
 export interface AccountSchema {
   /** bban */
   bban?: string;
@@ -76,263 +69,13 @@ export interface AccountSchema {
   pan?: string;
 }
 
-export enum AccountStatusEnum {
-  DISCOVERED = "DISCOVERED",
-  PROCESSING = "PROCESSING",
-  READY = "READY",
-  ERROR = "ERROR",
-  SUSPENDED = "SUSPENDED",
+/** AccountTransactionsSerializer. */
+export interface AccountTransactions {
+  /** transactions */
+  transactions: BankTransaction;
 }
 
-/**
- * * `AT` - Austria
- * * `BE` - Belgium
- * * `BG` - Bulgaria
- * * `HR` - Croatia
- * * `CY` - Cyprus
- * * `CZ` - Czechia
- * * `DK` - Denmark
- * * `EE` - Estonia
- * * `FI` - Finland
- * * `FR` - France
- * * `DE` - Germany
- * * `GR` - Greece
- * * `HU` - Hungary
- * * `IS` - Iceland
- * * `IE` - Ireland
- * * `IT` - Italy
- * * `LV` - Latvia
- * * `LI` - Liechtenstein
- * * `LT` - Lithuania
- * * `LU` - Luxembourg
- * * `MT` - Malta
- * * `NL` - Netherlands
- * * `NO` - Norway
- * * `PL` - Poland
- * * `PT` - Portugal
- * * `RO` - Romania
- * * `SK` - Slovakia
- * * `SI` - Slovenia
- * * `ES` - Spain
- * * `SE` - Sweden
- * * `GB` - United Kingdom
- * * `US` - United States of America
- */
-export enum AddressCountryEnum {
-  AT = "AT",
-  BE = "BE",
-  BG = "BG",
-  HR = "HR",
-  CY = "CY",
-  CZ = "CZ",
-  DK = "DK",
-  EE = "EE",
-  FI = "FI",
-  FR = "FR",
-  DE = "DE",
-  GR = "GR",
-  HU = "HU",
-  IS = "IS",
-  IE = "IE",
-  IT = "IT",
-  LV = "LV",
-  LI = "LI",
-  LT = "LT",
-  LU = "LU",
-  MT = "MT",
-  NL = "NL",
-  NO = "NO",
-  PL = "PL",
-  PT = "PT",
-  RO = "RO",
-  SK = "SK",
-  SI = "SI",
-  ES = "ES",
-  SE = "SE",
-  GB = "GB",
-  US = "US",
-}
-
-export interface BACS {
-  /** Creditor account */
-  creditor_object: CreditorAccountWrite;
-  /**
-   * Payment end to end identification.
-   * Payment Custom Payment ID
-   * @maxLength 35
-   */
-  custom_payment_id?: string;
-  /**
-   * Payment description
-   * @default "GOCARDLESS"
-   */
-  description?: string;
-  /** Instructed amount */
-  instructed_amount: InstructedAmount;
-  /**
-   * Link to initiate authorization with Institution
-   * @format uri
-   * @default "https://ob.nordigen.com/pis/start/1b1b329f-a654-4ed3-b244-8245de396aa8/{$INSTITUTION_ID}"
-   */
-  link?: string;
-  /** Payment ID */
-  payment_id?: string;
-  /**
-   * Payment product
-   *
-   * * `T2P` - target-2-payments
-   * * `SCT` - sepa-credit-transfers
-   * * `ISCT` - instant-sepa-credit-transfer
-   * * `CBCT` - cross-border-credit-transfers
-   * * `BACS` - Back Payment Scheme
-   * * `CHAPS` - CHAPS Payment Scheme
-   * * `FPS` - Faster Payment Scheme
-   * * `SWIFT` - Swift Payment Service
-   * * `BT` - Balance Transfer
-   * * `MT` - Money Transfer
-   * * `DCT` - domestic-credit-transfer
-   * * `IDCT` - instant-domestic-credit-transfer
-   * @default "ISCT"
-   */
-  payment_product?: PaymentProductEnum;
-  /**
-   * Payment end to end identification
-   *
-   * * `INIT` - Initiated. Payment has been initiated.
-   * * `ERRE` - ExecutionError. We experienced error on payment execution.
-   * * `ERRS` - StatusError. We experienced error retrieving payment status. Try again.
-   * * `ACCC` - AcceptedSettlementCompleted. Settlement on the creditor's account has been completed
-   * * `ACCP` - AcceptedCustomerProfile. Preceding check of technical validation was successful. Customer profile check was successful
-   * * `ACSC` - AcceptedSettlementCompleted. Settlement on the debtor’s account has been completed
-   * * `ACSP` - AcceptedSettlementInProcess. All preceding checks such as technical validation and customer profile were successful and therefore the payment initiation has been accepted for execution
-   * * `ACTC` - AcceptedTechnicalValidation. Authentication and syntactical and semantical validation are successful
-   * * `ACWC` - AcceptedWithChange. Instruction is accepted but a change will be made, such as date or remittance not sent
-   * * `ACWP` - AcceptedWithoutPosting. Payment instruction included in the credit transfer is accepted without being posted to the creditor customer’s account
-   * * `RCVD` - Received. Payment initiation has been received by the receiving agent
-   * * `PDNG` - Pending. Payment initiation or individual transaction included in the payment initiation is pending. Further checks and status update will be performed
-   * * `RJCT` - Rejected. Payment initiation or individual transaction included in the payment initiation has been rejected.
-   * * `CANC` - Cancelled. Payment initiation has been cancelled before execution
-   * * `ACFC` - AcceptedFundsChecked. Pre-ceeding check of technical validation and customer profile was successful and an automatic funds check was positive
-   * * `PATC` - PartiallyAcceptedTechnicalCorrect. The payment initiation needs multiple authentications, where some but not yet all have been performed
-   * * `PART` - PartiallyAccepted. A number of transactions have been accepted, whereas another number of transactions have not yet achieved 'accepted' status
-   */
-  payment_status?: PaymentStatusEnum;
-  /**
-   * Payment Type
-   *
-   * * `single-payment` - payment
-   * * `bulk-payment` - bulk-payments
-   * * `periodic-payment` - periodic-payments
-   */
-  payment_type?: PaymentTypeEnum;
-  /**
-   * Redirect URL to your application after payment is done
-   * @format uri
-   * @maxLength 1024
-   */
-  redirect: string | null;
-  /**
-   * Payment Execution date (for periodic payments)
-   * @format date
-   */
-  requested_execution_date?: string;
-  /**
-   * Indicates whether payment should be submitted separately
-   * @default false
-   */
-  submit_payment?: boolean;
-}
-
-export interface BT {
-  /** Creditor account */
-  creditor_object: CreditorAccountWrite;
-  /**
-   * Payment end to end identification.
-   * Payment Custom Payment ID
-   * @maxLength 35
-   */
-  custom_payment_id?: string;
-  /**
-   * Payment description
-   * @default "GOCARDLESS"
-   */
-  description?: string;
-  /** Instructed amount */
-  instructed_amount: InstructedAmount;
-  /**
-   * Link to initiate authorization with Institution
-   * @format uri
-   * @default "https://ob.nordigen.com/pis/start/1b1b329f-a654-4ed3-b244-8245de396aa8/{$INSTITUTION_ID}"
-   */
-  link?: string;
-  /** Payment ID */
-  payment_id?: string;
-  /**
-   * Payment product
-   *
-   * * `T2P` - target-2-payments
-   * * `SCT` - sepa-credit-transfers
-   * * `ISCT` - instant-sepa-credit-transfer
-   * * `CBCT` - cross-border-credit-transfers
-   * * `BACS` - Back Payment Scheme
-   * * `CHAPS` - CHAPS Payment Scheme
-   * * `FPS` - Faster Payment Scheme
-   * * `SWIFT` - Swift Payment Service
-   * * `BT` - Balance Transfer
-   * * `MT` - Money Transfer
-   * * `DCT` - domestic-credit-transfer
-   * * `IDCT` - instant-domestic-credit-transfer
-   * @default "ISCT"
-   */
-  payment_product?: PaymentProductEnum;
-  /**
-   * Payment end to end identification
-   *
-   * * `INIT` - Initiated. Payment has been initiated.
-   * * `ERRE` - ExecutionError. We experienced error on payment execution.
-   * * `ERRS` - StatusError. We experienced error retrieving payment status. Try again.
-   * * `ACCC` - AcceptedSettlementCompleted. Settlement on the creditor's account has been completed
-   * * `ACCP` - AcceptedCustomerProfile. Preceding check of technical validation was successful. Customer profile check was successful
-   * * `ACSC` - AcceptedSettlementCompleted. Settlement on the debtor’s account has been completed
-   * * `ACSP` - AcceptedSettlementInProcess. All preceding checks such as technical validation and customer profile were successful and therefore the payment initiation has been accepted for execution
-   * * `ACTC` - AcceptedTechnicalValidation. Authentication and syntactical and semantical validation are successful
-   * * `ACWC` - AcceptedWithChange. Instruction is accepted but a change will be made, such as date or remittance not sent
-   * * `ACWP` - AcceptedWithoutPosting. Payment instruction included in the credit transfer is accepted without being posted to the creditor customer’s account
-   * * `RCVD` - Received. Payment initiation has been received by the receiving agent
-   * * `PDNG` - Pending. Payment initiation or individual transaction included in the payment initiation is pending. Further checks and status update will be performed
-   * * `RJCT` - Rejected. Payment initiation or individual transaction included in the payment initiation has been rejected.
-   * * `CANC` - Cancelled. Payment initiation has been cancelled before execution
-   * * `ACFC` - AcceptedFundsChecked. Pre-ceeding check of technical validation and customer profile was successful and an automatic funds check was positive
-   * * `PATC` - PartiallyAcceptedTechnicalCorrect. The payment initiation needs multiple authentications, where some but not yet all have been performed
-   * * `PART` - PartiallyAccepted. A number of transactions have been accepted, whereas another number of transactions have not yet achieved 'accepted' status
-   */
-  payment_status?: PaymentStatusEnum;
-  /**
-   * Payment Type
-   *
-   * * `single-payment` - payment
-   * * `bulk-payment` - bulk-payments
-   * * `periodic-payment` - periodic-payments
-   */
-  payment_type?: PaymentTypeEnum;
-  /**
-   * Redirect URL to your application after payment is done
-   * @format uri
-   * @maxLength 1024
-   */
-  redirect: string | null;
-  /**
-   * Payment Execution date (for periodic payments)
-   * @format date
-   */
-  requested_execution_date?: string;
-  /**
-   * Indicates whether payment should be submitted separately
-   * @default false
-   */
-  submit_payment?: boolean;
-}
-
+/** BalanceAmountSchema. */
 export interface BalanceAmountSchema {
   /** amount */
   amount: string;
@@ -340,6 +83,7 @@ export interface BalanceAmountSchema {
   currency: string;
 }
 
+/** BalanceSchema. */
 export interface BalanceSchema {
   /** balanceAmount */
   balanceAmount: BalanceAmountSchema;
@@ -355,258 +99,10 @@ export interface BalanceSchema {
   referenceDate?: string;
 }
 
-export interface BankTransactionStatusSchema {
+/** BankTransactionSerializer. */
+export interface BankTransaction {
   booked: TransactionSchema[];
   pending?: TransactionSchema[];
-}
-
-export interface CBCT {
-  /** Creditor account */
-  creditor_object: CreditorAccountWrite;
-  /**
-   * Payment end to end identification.
-   * Payment Custom Payment ID
-   * @maxLength 35
-   */
-  custom_payment_id?: string;
-  /**
-   * Payment description
-   * @default "GOCARDLESS"
-   */
-  description?: string;
-  /** Instructed amount */
-  instructed_amount: InstructedAmount;
-  /**
-   * Link to initiate authorization with Institution
-   * @format uri
-   * @default "https://ob.nordigen.com/pis/start/1b1b329f-a654-4ed3-b244-8245de396aa8/{$INSTITUTION_ID}"
-   */
-  link?: string;
-  /** Payment ID */
-  payment_id?: string;
-  /**
-   * Payment product
-   *
-   * * `T2P` - target-2-payments
-   * * `SCT` - sepa-credit-transfers
-   * * `ISCT` - instant-sepa-credit-transfer
-   * * `CBCT` - cross-border-credit-transfers
-   * * `BACS` - Back Payment Scheme
-   * * `CHAPS` - CHAPS Payment Scheme
-   * * `FPS` - Faster Payment Scheme
-   * * `SWIFT` - Swift Payment Service
-   * * `BT` - Balance Transfer
-   * * `MT` - Money Transfer
-   * * `DCT` - domestic-credit-transfer
-   * * `IDCT` - instant-domestic-credit-transfer
-   * @default "ISCT"
-   */
-  payment_product?: PaymentProductEnum;
-  /**
-   * Payment end to end identification
-   *
-   * * `INIT` - Initiated. Payment has been initiated.
-   * * `ERRE` - ExecutionError. We experienced error on payment execution.
-   * * `ERRS` - StatusError. We experienced error retrieving payment status. Try again.
-   * * `ACCC` - AcceptedSettlementCompleted. Settlement on the creditor's account has been completed
-   * * `ACCP` - AcceptedCustomerProfile. Preceding check of technical validation was successful. Customer profile check was successful
-   * * `ACSC` - AcceptedSettlementCompleted. Settlement on the debtor’s account has been completed
-   * * `ACSP` - AcceptedSettlementInProcess. All preceding checks such as technical validation and customer profile were successful and therefore the payment initiation has been accepted for execution
-   * * `ACTC` - AcceptedTechnicalValidation. Authentication and syntactical and semantical validation are successful
-   * * `ACWC` - AcceptedWithChange. Instruction is accepted but a change will be made, such as date or remittance not sent
-   * * `ACWP` - AcceptedWithoutPosting. Payment instruction included in the credit transfer is accepted without being posted to the creditor customer’s account
-   * * `RCVD` - Received. Payment initiation has been received by the receiving agent
-   * * `PDNG` - Pending. Payment initiation or individual transaction included in the payment initiation is pending. Further checks and status update will be performed
-   * * `RJCT` - Rejected. Payment initiation or individual transaction included in the payment initiation has been rejected.
-   * * `CANC` - Cancelled. Payment initiation has been cancelled before execution
-   * * `ACFC` - AcceptedFundsChecked. Pre-ceeding check of technical validation and customer profile was successful and an automatic funds check was positive
-   * * `PATC` - PartiallyAcceptedTechnicalCorrect. The payment initiation needs multiple authentications, where some but not yet all have been performed
-   * * `PART` - PartiallyAccepted. A number of transactions have been accepted, whereas another number of transactions have not yet achieved 'accepted' status
-   */
-  payment_status?: PaymentStatusEnum;
-  /**
-   * Payment Type
-   *
-   * * `single-payment` - payment
-   * * `bulk-payment` - bulk-payments
-   * * `periodic-payment` - periodic-payments
-   */
-  payment_type?: PaymentTypeEnum;
-  /**
-   * Redirect URL to your application after payment is done
-   * @format uri
-   * @maxLength 1024
-   */
-  redirect: string | null;
-  /**
-   * Payment Execution date (for periodic payments)
-   * @format date
-   */
-  requested_execution_date?: string;
-  /**
-   * Indicates whether payment should be submitted separately
-   * @default false
-   */
-  submit_payment?: boolean;
-}
-
-export interface CHAPS {
-  /** Creditor account */
-  creditor_object: CreditorAccountWrite;
-  /**
-   * Payment end to end identification.
-   * Payment Custom Payment ID
-   * @maxLength 35
-   */
-  custom_payment_id?: string;
-  /**
-   * Payment description
-   * @default "GOCARDLESS"
-   */
-  description?: string;
-  /** Instructed amount */
-  instructed_amount: InstructedAmount;
-  /**
-   * Link to initiate authorization with Institution
-   * @format uri
-   * @default "https://ob.nordigen.com/pis/start/1b1b329f-a654-4ed3-b244-8245de396aa8/{$INSTITUTION_ID}"
-   */
-  link?: string;
-  /** Payment ID */
-  payment_id?: string;
-  /**
-   * Payment product
-   *
-   * * `T2P` - target-2-payments
-   * * `SCT` - sepa-credit-transfers
-   * * `ISCT` - instant-sepa-credit-transfer
-   * * `CBCT` - cross-border-credit-transfers
-   * * `BACS` - Back Payment Scheme
-   * * `CHAPS` - CHAPS Payment Scheme
-   * * `FPS` - Faster Payment Scheme
-   * * `SWIFT` - Swift Payment Service
-   * * `BT` - Balance Transfer
-   * * `MT` - Money Transfer
-   * * `DCT` - domestic-credit-transfer
-   * * `IDCT` - instant-domestic-credit-transfer
-   * @default "ISCT"
-   */
-  payment_product?: PaymentProductEnum;
-  /**
-   * Payment end to end identification
-   *
-   * * `INIT` - Initiated. Payment has been initiated.
-   * * `ERRE` - ExecutionError. We experienced error on payment execution.
-   * * `ERRS` - StatusError. We experienced error retrieving payment status. Try again.
-   * * `ACCC` - AcceptedSettlementCompleted. Settlement on the creditor's account has been completed
-   * * `ACCP` - AcceptedCustomerProfile. Preceding check of technical validation was successful. Customer profile check was successful
-   * * `ACSC` - AcceptedSettlementCompleted. Settlement on the debtor’s account has been completed
-   * * `ACSP` - AcceptedSettlementInProcess. All preceding checks such as technical validation and customer profile were successful and therefore the payment initiation has been accepted for execution
-   * * `ACTC` - AcceptedTechnicalValidation. Authentication and syntactical and semantical validation are successful
-   * * `ACWC` - AcceptedWithChange. Instruction is accepted but a change will be made, such as date or remittance not sent
-   * * `ACWP` - AcceptedWithoutPosting. Payment instruction included in the credit transfer is accepted without being posted to the creditor customer’s account
-   * * `RCVD` - Received. Payment initiation has been received by the receiving agent
-   * * `PDNG` - Pending. Payment initiation or individual transaction included in the payment initiation is pending. Further checks and status update will be performed
-   * * `RJCT` - Rejected. Payment initiation or individual transaction included in the payment initiation has been rejected.
-   * * `CANC` - Cancelled. Payment initiation has been cancelled before execution
-   * * `ACFC` - AcceptedFundsChecked. Pre-ceeding check of technical validation and customer profile was successful and an automatic funds check was positive
-   * * `PATC` - PartiallyAcceptedTechnicalCorrect. The payment initiation needs multiple authentications, where some but not yet all have been performed
-   * * `PART` - PartiallyAccepted. A number of transactions have been accepted, whereas another number of transactions have not yet achieved 'accepted' status
-   */
-  payment_status?: PaymentStatusEnum;
-  /**
-   * Payment Type
-   *
-   * * `single-payment` - payment
-   * * `bulk-payment` - bulk-payments
-   * * `periodic-payment` - periodic-payments
-   */
-  payment_type?: PaymentTypeEnum;
-  /**
-   * Redirect URL to your application after payment is done
-   * @format uri
-   * @maxLength 1024
-   */
-  redirect: string | null;
-  /**
-   * Payment Execution date (for periodic payments)
-   * @format date
-   */
-  requested_execution_date?: string;
-  /**
-   * Indicates whether payment should be submitted separately
-   * @default false
-   */
-  submit_payment?: boolean;
-}
-
-/**
- * * `AT` - Austria
- * * `BE` - Belgium
- * * `BG` - Bulgaria
- * * `HR` - Croatia
- * * `CY` - Cyprus
- * * `CZ` - Czechia
- * * `DK` - Denmark
- * * `EE` - Estonia
- * * `FI` - Finland
- * * `FR` - France
- * * `DE` - Germany
- * * `GR` - Greece
- * * `HU` - Hungary
- * * `IS` - Iceland
- * * `IE` - Ireland
- * * `IT` - Italy
- * * `LV` - Latvia
- * * `LI` - Liechtenstein
- * * `LT` - Lithuania
- * * `LU` - Luxembourg
- * * `MT` - Malta
- * * `NL` - Netherlands
- * * `NO` - Norway
- * * `PL` - Poland
- * * `PT` - Portugal
- * * `RO` - Romania
- * * `SK` - Slovakia
- * * `SI` - Slovenia
- * * `ES` - Spain
- * * `SE` - Sweden
- * * `GB` - United Kingdom
- * * `US` - United States of America
- */
-export enum CountryEnum {
-  AT = "AT",
-  BE = "BE",
-  BG = "BG",
-  HR = "HR",
-  CY = "CY",
-  CZ = "CZ",
-  DK = "DK",
-  EE = "EE",
-  FI = "FI",
-  FR = "FR",
-  DE = "DE",
-  GR = "GR",
-  HU = "HU",
-  IS = "IS",
-  IE = "IE",
-  IT = "IT",
-  LV = "LV",
-  LI = "LI",
-  LT = "LT",
-  LU = "LU",
-  MT = "MT",
-  NL = "NL",
-  NO = "NO",
-  PL = "PL",
-  PT = "PT",
-  RO = "RO",
-  SK = "SK",
-  SI = "SI",
-  ES = "ES",
-  SE = "SE",
-  GB = "GB",
-  US = "US",
 }
 
 export type CreateEuaData = EndUserAgreement;
@@ -617,240 +113,7 @@ export type CreateRequisitionData = SpectacularRequisition;
 
 export type CreateRequisitionError = ErrorResponse;
 
-/** Creditor account read serializer. */
-export interface CreditorAccount {
-  /**
-   * Creditor account type identifier
-   * @maxLength 128
-   */
-  account: string;
-  /**
-   * Creditor account address country
-   * @maxLength 2
-   */
-  address_country?: string;
-  /**
-   * Creditor account currency
-   * @maxLength 3
-   */
-  currency: string;
-  /**
-   * Unique entry ID
-   * @format uuid
-   */
-  id?: string;
-  /**
-   * Creditor account name
-   * @maxLength 70
-   */
-  name: string;
-  /**
-   * Creditor account type
-   *
-   * * `IBAN` - IBAN
-   * * `SCAN` - SortCodeAccountNumber
-   * * `BBAN` - BBAN
-   * @default "IBAN"
-   */
-  type?: TypeEnum;
-}
-
-/** Creditor account write serializer. */
-export interface CreditorAccountWrite {
-  /**
-   * Creditor account type identifier
-   * @maxLength 128
-   */
-  account: string;
-  /**
-   * Creditor account address city
-   * @maxLength 140
-   */
-  address_city?: string;
-  /**
-   * * `AT` - Austria
-   * * `BE` - Belgium
-   * * `BG` - Bulgaria
-   * * `HR` - Croatia
-   * * `CY` - Cyprus
-   * * `CZ` - Czechia
-   * * `DK` - Denmark
-   * * `EE` - Estonia
-   * * `FI` - Finland
-   * * `FR` - France
-   * * `DE` - Germany
-   * * `GR` - Greece
-   * * `HU` - Hungary
-   * * `IS` - Iceland
-   * * `IE` - Ireland
-   * * `IT` - Italy
-   * * `LV` - Latvia
-   * * `LI` - Liechtenstein
-   * * `LT` - Lithuania
-   * * `LU` - Luxembourg
-   * * `MT` - Malta
-   * * `NL` - Netherlands
-   * * `NO` - Norway
-   * * `PL` - Poland
-   * * `PT` - Portugal
-   * * `RO` - Romania
-   * * `SK` - Slovakia
-   * * `SI` - Slovenia
-   * * `ES` - Spain
-   * * `SE` - Sweden
-   * * `GB` - United Kingdom
-   * * `US` - United States of America
-   */
-  address_country?: AddressCountryEnum;
-  /**
-   * Creditor account address street
-   * @maxLength 140
-   */
-  address_street?: string;
-  /**
-   * Creditor account BICFI Identifier
-   * @maxLength 128
-   */
-  agent?: string;
-  /**
-   * Creditor account agent name
-   * @maxLength 140
-   */
-  agent_name?: string;
-  /**
-   * Creditor account currency
-   * @maxLength 3
-   */
-  currency: string;
-  /**
-   * Unique entry ID
-   * @format uuid
-   */
-  id?: string;
-  /** an Institution ID for this CreditorAccount */
-  institution_id?: string;
-  /**
-   * Creditor account name
-   * @maxLength 70
-   */
-  name: string;
-  /**
-   * Creditor account address post code
-   * @maxLength 30
-   */
-  post_code?: string;
-  /**
-   * Creditor account type
-   *
-   * * `IBAN` - IBAN
-   * * `SCAN` - SortCodeAccountNumber
-   * * `BBAN` - BBAN
-   * @default "IBAN"
-   */
-  type?: TypeEnum;
-}
-
-/** Creditor account write serializer. */
-export interface CreditorAccountWriteRequest {
-  /**
-   * Creditor account type identifier
-   * @minLength 1
-   * @maxLength 128
-   */
-  account: string;
-  /**
-   * Creditor account address city
-   * @minLength 1
-   * @maxLength 140
-   */
-  address_city?: string;
-  /**
-   * * `AT` - Austria
-   * * `BE` - Belgium
-   * * `BG` - Bulgaria
-   * * `HR` - Croatia
-   * * `CY` - Cyprus
-   * * `CZ` - Czechia
-   * * `DK` - Denmark
-   * * `EE` - Estonia
-   * * `FI` - Finland
-   * * `FR` - France
-   * * `DE` - Germany
-   * * `GR` - Greece
-   * * `HU` - Hungary
-   * * `IS` - Iceland
-   * * `IE` - Ireland
-   * * `IT` - Italy
-   * * `LV` - Latvia
-   * * `LI` - Liechtenstein
-   * * `LT` - Lithuania
-   * * `LU` - Luxembourg
-   * * `MT` - Malta
-   * * `NL` - Netherlands
-   * * `NO` - Norway
-   * * `PL` - Poland
-   * * `PT` - Portugal
-   * * `RO` - Romania
-   * * `SK` - Slovakia
-   * * `SI` - Slovenia
-   * * `ES` - Spain
-   * * `SE` - Sweden
-   * * `GB` - United Kingdom
-   * * `US` - United States of America
-   */
-  address_country?: AddressCountryEnum;
-  /**
-   * Creditor account address street
-   * @minLength 1
-   * @maxLength 140
-   */
-  address_street?: string;
-  /**
-   * Creditor account BICFI Identifier
-   * @minLength 1
-   * @maxLength 128
-   */
-  agent?: string;
-  /**
-   * Creditor account agent name
-   * @minLength 1
-   * @maxLength 140
-   */
-  agent_name?: string;
-  /**
-   * Creditor account currency
-   * @minLength 1
-   * @maxLength 3
-   */
-  currency: string;
-  /**
-   * an Institution ID for this CreditorAccount
-   * @minLength 1
-   */
-  institution_id?: string;
-  /**
-   * Creditor account name
-   * @minLength 1
-   * @maxLength 70
-   */
-  name: string;
-  /**
-   * Creditor account address post code
-   * @minLength 1
-   * @maxLength 30
-   */
-  post_code?: string;
-  /**
-   * Creditor account type
-   *
-   * * `IBAN` - IBAN
-   * * `SCAN` - SortCodeAccountNumber
-   * * `BBAN` - BBAN
-   * @default "IBAN"
-   */
-  type?: TypeEnum;
-}
-
+/** CurrencyExchangeSchema. */
 export interface CurrencyExchangeSchema {
   /** contractIdentification */
   contractIdentification?: string;
@@ -866,214 +129,15 @@ export interface CurrencyExchangeSchema {
   unitCurrency?: string;
 }
 
-export interface DCT {
-  /** Creditor account */
-  creditor_object: CreditorAccountWrite;
-  /**
-   * Payment end to end identification.
-   * Payment Custom Payment ID
-   * @maxLength 35
-   */
-  custom_payment_id?: string;
-  /**
-   * Payment description
-   * @default "GOCARDLESS"
-   */
-  description?: string;
-  /** Instructed amount */
-  instructed_amount: InstructedAmount;
-  /**
-   * Link to initiate authorization with Institution
-   * @format uri
-   * @default "https://ob.nordigen.com/pis/start/1b1b329f-a654-4ed3-b244-8245de396aa8/{$INSTITUTION_ID}"
-   */
-  link?: string;
-  /** Payment ID */
-  payment_id?: string;
-  /**
-   * Payment product
-   *
-   * * `T2P` - target-2-payments
-   * * `SCT` - sepa-credit-transfers
-   * * `ISCT` - instant-sepa-credit-transfer
-   * * `CBCT` - cross-border-credit-transfers
-   * * `BACS` - Back Payment Scheme
-   * * `CHAPS` - CHAPS Payment Scheme
-   * * `FPS` - Faster Payment Scheme
-   * * `SWIFT` - Swift Payment Service
-   * * `BT` - Balance Transfer
-   * * `MT` - Money Transfer
-   * * `DCT` - domestic-credit-transfer
-   * * `IDCT` - instant-domestic-credit-transfer
-   * @default "ISCT"
-   */
-  payment_product?: PaymentProductEnum;
-  /**
-   * Payment end to end identification
-   *
-   * * `INIT` - Initiated. Payment has been initiated.
-   * * `ERRE` - ExecutionError. We experienced error on payment execution.
-   * * `ERRS` - StatusError. We experienced error retrieving payment status. Try again.
-   * * `ACCC` - AcceptedSettlementCompleted. Settlement on the creditor's account has been completed
-   * * `ACCP` - AcceptedCustomerProfile. Preceding check of technical validation was successful. Customer profile check was successful
-   * * `ACSC` - AcceptedSettlementCompleted. Settlement on the debtor’s account has been completed
-   * * `ACSP` - AcceptedSettlementInProcess. All preceding checks such as technical validation and customer profile were successful and therefore the payment initiation has been accepted for execution
-   * * `ACTC` - AcceptedTechnicalValidation. Authentication and syntactical and semantical validation are successful
-   * * `ACWC` - AcceptedWithChange. Instruction is accepted but a change will be made, such as date or remittance not sent
-   * * `ACWP` - AcceptedWithoutPosting. Payment instruction included in the credit transfer is accepted without being posted to the creditor customer’s account
-   * * `RCVD` - Received. Payment initiation has been received by the receiving agent
-   * * `PDNG` - Pending. Payment initiation or individual transaction included in the payment initiation is pending. Further checks and status update will be performed
-   * * `RJCT` - Rejected. Payment initiation or individual transaction included in the payment initiation has been rejected.
-   * * `CANC` - Cancelled. Payment initiation has been cancelled before execution
-   * * `ACFC` - AcceptedFundsChecked. Pre-ceeding check of technical validation and customer profile was successful and an automatic funds check was positive
-   * * `PATC` - PartiallyAcceptedTechnicalCorrect. The payment initiation needs multiple authentications, where some but not yet all have been performed
-   * * `PART` - PartiallyAccepted. A number of transactions have been accepted, whereas another number of transactions have not yet achieved 'accepted' status
-   */
-  payment_status?: PaymentStatusEnum;
-  /**
-   * Payment Type
-   *
-   * * `single-payment` - payment
-   * * `bulk-payment` - bulk-payments
-   * * `periodic-payment` - periodic-payments
-   */
-  payment_type?: PaymentTypeEnum;
-  /**
-   * Redirect URL to your application after payment is done
-   * @format uri
-   * @maxLength 1024
-   */
-  redirect: string | null;
-  /**
-   * Payment Execution date (for periodic payments)
-   * @format date
-   */
-  requested_execution_date?: string;
-  /**
-   * Indicates whether payment should be submitted separately
-   * @default false
-   */
-  submit_payment?: boolean;
-}
-
-/** Debtor account write serializer. */
-export interface DebtorAccountWrite {
-  /**
-   * Debtor account type identifier
-   * @maxLength 128
-   */
-  account?: string;
-  /**
-   * Debtor account address country
-   * @maxLength 2
-   */
-  address_country?: string;
-  /**
-   * Debtor account address street
-   * @maxLength 140
-   */
-  address_street?: string;
-  /**
-   * Creditor account BICFI Identifier
-   * @maxLength 128
-   */
-  agent?: string;
-  /**
-   * Debtor account currency
-   * @maxLength 3
-   */
-  currency?: string;
-  /**
-   * Debtor account name
-   * @maxLength 70
-   */
-  name: string;
-  /**
-   * Debtor account post code
-   * @maxLength 30
-   */
-  post_code?: string;
-  /**
-   * Debtor account type
-   *
-   * * `IBAN` - IBAN
-   * * `SCAN` - SortCodeAccountNumber
-   * * `BBAN` - BBAN
-   * @default "IBAN"
-   */
-  type?: TypeEnum;
-  /**
-   * Debtor account type identifier
-   * @maxLength 128
-   */
-  type_number?: string;
-}
-
-/** Debtor account write serializer. */
-export interface DebtorAccountWriteRequest {
-  /**
-   * Debtor account type identifier
-   * @minLength 1
-   * @maxLength 128
-   */
-  account?: string;
-  /**
-   * Debtor account address country
-   * @minLength 1
-   * @maxLength 2
-   */
-  address_country?: string;
-  /**
-   * Debtor account address street
-   * @minLength 1
-   * @maxLength 140
-   */
-  address_street?: string;
-  /**
-   * Creditor account BICFI Identifier
-   * @minLength 1
-   * @maxLength 128
-   */
-  agent?: string;
-  /**
-   * Debtor account currency
-   * @minLength 1
-   * @maxLength 3
-   */
-  currency?: string;
-  /**
-   * Debtor account name
-   * @minLength 1
-   * @maxLength 70
-   */
-  name: string;
-  /**
-   * Debtor account post code
-   * @minLength 1
-   * @maxLength 30
-   */
-  post_code?: string;
-  /**
-   * Debtor account type
-   *
-   * * `IBAN` - IBAN
-   * * `SCAN` - SortCodeAccountNumber
-   * * `BBAN` - BBAN
-   * @default "IBAN"
-   */
-  type?: TypeEnum;
-  /**
-   * Debtor account type identifier
-   * @minLength 1
-   * @maxLength 128
-   */
-  type_number?: string;
-}
+export type DeleteEuaByIdData = SuccessfulDeleteResponse;
 
 export type DeleteEuaByIdError = ErrorResponse;
 
+export type DeleteRequisitionByIdData = SuccessfulDeleteResponse;
+
 export type DeleteRequisitionByIdError = ErrorResponse;
 
+/** DetailSchema. */
 export interface DetailSchema {
   /** bban */
   bban?: string;
@@ -1107,6 +171,8 @@ export interface DetailSchema {
   product?: string;
   /** resourceId */
   resourceId?: string;
+  /** SortCodeAccountNumber returned by some UK banks (6 digit Sort Code and 8 digit Account Number) */
+  scan?: string;
   /** status */
   status?: string;
   /** usage */
@@ -1211,324 +277,9 @@ export interface ErrorResponse {
   type?: string;
 }
 
-/**
- * * `following` - following
- * * `preceding` - preceding
- */
-export enum ExecutionRuleEnum {
-  Following = "following",
-  Preceding = "preceding",
-}
-
-export interface FPS {
-  /** Creditor account */
-  creditor_object: CreditorAccountWrite;
-  /**
-   * Payment end to end identification.
-   * Payment Custom Payment ID
-   * @maxLength 35
-   */
-  custom_payment_id?: string;
-  /**
-   * Payment description
-   * @default "GOCARDLESS"
-   */
-  description?: string;
-  /** Instructed amount */
-  instructed_amount: InstructedAmount;
-  /**
-   * Link to initiate authorization with Institution
-   * @format uri
-   * @default "https://ob.nordigen.com/pis/start/1b1b329f-a654-4ed3-b244-8245de396aa8/{$INSTITUTION_ID}"
-   */
-  link?: string;
-  /** Payment ID */
-  payment_id?: string;
-  /**
-   * Payment product
-   *
-   * * `T2P` - target-2-payments
-   * * `SCT` - sepa-credit-transfers
-   * * `ISCT` - instant-sepa-credit-transfer
-   * * `CBCT` - cross-border-credit-transfers
-   * * `BACS` - Back Payment Scheme
-   * * `CHAPS` - CHAPS Payment Scheme
-   * * `FPS` - Faster Payment Scheme
-   * * `SWIFT` - Swift Payment Service
-   * * `BT` - Balance Transfer
-   * * `MT` - Money Transfer
-   * * `DCT` - domestic-credit-transfer
-   * * `IDCT` - instant-domestic-credit-transfer
-   * @default "ISCT"
-   */
-  payment_product?: PaymentProductEnum;
-  /**
-   * Payment end to end identification
-   *
-   * * `INIT` - Initiated. Payment has been initiated.
-   * * `ERRE` - ExecutionError. We experienced error on payment execution.
-   * * `ERRS` - StatusError. We experienced error retrieving payment status. Try again.
-   * * `ACCC` - AcceptedSettlementCompleted. Settlement on the creditor's account has been completed
-   * * `ACCP` - AcceptedCustomerProfile. Preceding check of technical validation was successful. Customer profile check was successful
-   * * `ACSC` - AcceptedSettlementCompleted. Settlement on the debtor’s account has been completed
-   * * `ACSP` - AcceptedSettlementInProcess. All preceding checks such as technical validation and customer profile were successful and therefore the payment initiation has been accepted for execution
-   * * `ACTC` - AcceptedTechnicalValidation. Authentication and syntactical and semantical validation are successful
-   * * `ACWC` - AcceptedWithChange. Instruction is accepted but a change will be made, such as date or remittance not sent
-   * * `ACWP` - AcceptedWithoutPosting. Payment instruction included in the credit transfer is accepted without being posted to the creditor customer’s account
-   * * `RCVD` - Received. Payment initiation has been received by the receiving agent
-   * * `PDNG` - Pending. Payment initiation or individual transaction included in the payment initiation is pending. Further checks and status update will be performed
-   * * `RJCT` - Rejected. Payment initiation or individual transaction included in the payment initiation has been rejected.
-   * * `CANC` - Cancelled. Payment initiation has been cancelled before execution
-   * * `ACFC` - AcceptedFundsChecked. Pre-ceeding check of technical validation and customer profile was successful and an automatic funds check was positive
-   * * `PATC` - PartiallyAcceptedTechnicalCorrect. The payment initiation needs multiple authentications, where some but not yet all have been performed
-   * * `PART` - PartiallyAccepted. A number of transactions have been accepted, whereas another number of transactions have not yet achieved 'accepted' status
-   */
-  payment_status?: PaymentStatusEnum;
-  /**
-   * Payment Type
-   *
-   * * `single-payment` - payment
-   * * `bulk-payment` - bulk-payments
-   * * `periodic-payment` - periodic-payments
-   */
-  payment_type?: PaymentTypeEnum;
-  /**
-   * Redirect URL to your application after payment is done
-   * @format uri
-   * @maxLength 1024
-   */
-  redirect: string | null;
-  /**
-   * Payment Execution date (for periodic payments)
-   * @format date
-   */
-  requested_execution_date?: string;
-  /**
-   * Indicates whether payment should be submitted separately
-   * @default false
-   */
-  submit_payment?: boolean;
-}
-
-/**
- * * `Daily` - Daily
- * * `Weekly` - Weekly
- * * `EveryTwoWeeks` - EveryTwoWeeks
- * * `Monthly` - Monthly
- * * `EveryTwoMonths` - EveryTwoMonths
- * * `Quarterly` - Quarterly
- * * `SemiAnnual` - SemiAnnual
- * * `Annual` - Annual
- * * `MonthlyVariable` - MonthlyVariable
- */
-export enum FrequencyEnum {
-  Daily = "Daily",
-  Weekly = "Weekly",
-  EveryTwoWeeks = "EveryTwoWeeks",
-  Monthly = "Monthly",
-  EveryTwoMonths = "EveryTwoMonths",
-  Quarterly = "Quarterly",
-  SemiAnnual = "SemiAnnual",
-  Annual = "Annual",
-  MonthlyVariable = "MonthlyVariable",
-}
-
 export type GetANewAccessTokenData = SpectacularJWTRefresh;
 
 export type GetANewAccessTokenError = ErrorResponse;
-
-export interface IDCT {
-  /** Creditor account */
-  creditor_object: CreditorAccountWrite;
-  /**
-   * Payment end to end identification.
-   * Payment Custom Payment ID
-   * @maxLength 35
-   */
-  custom_payment_id?: string;
-  /**
-   * Payment description
-   * @default "GOCARDLESS"
-   */
-  description?: string;
-  /** Instructed amount */
-  instructed_amount: InstructedAmount;
-  /**
-   * Link to initiate authorization with Institution
-   * @format uri
-   * @default "https://ob.nordigen.com/pis/start/1b1b329f-a654-4ed3-b244-8245de396aa8/{$INSTITUTION_ID}"
-   */
-  link?: string;
-  /** Payment ID */
-  payment_id?: string;
-  /**
-   * Payment product
-   *
-   * * `T2P` - target-2-payments
-   * * `SCT` - sepa-credit-transfers
-   * * `ISCT` - instant-sepa-credit-transfer
-   * * `CBCT` - cross-border-credit-transfers
-   * * `BACS` - Back Payment Scheme
-   * * `CHAPS` - CHAPS Payment Scheme
-   * * `FPS` - Faster Payment Scheme
-   * * `SWIFT` - Swift Payment Service
-   * * `BT` - Balance Transfer
-   * * `MT` - Money Transfer
-   * * `DCT` - domestic-credit-transfer
-   * * `IDCT` - instant-domestic-credit-transfer
-   * @default "ISCT"
-   */
-  payment_product?: PaymentProductEnum;
-  /**
-   * Payment end to end identification
-   *
-   * * `INIT` - Initiated. Payment has been initiated.
-   * * `ERRE` - ExecutionError. We experienced error on payment execution.
-   * * `ERRS` - StatusError. We experienced error retrieving payment status. Try again.
-   * * `ACCC` - AcceptedSettlementCompleted. Settlement on the creditor's account has been completed
-   * * `ACCP` - AcceptedCustomerProfile. Preceding check of technical validation was successful. Customer profile check was successful
-   * * `ACSC` - AcceptedSettlementCompleted. Settlement on the debtor’s account has been completed
-   * * `ACSP` - AcceptedSettlementInProcess. All preceding checks such as technical validation and customer profile were successful and therefore the payment initiation has been accepted for execution
-   * * `ACTC` - AcceptedTechnicalValidation. Authentication and syntactical and semantical validation are successful
-   * * `ACWC` - AcceptedWithChange. Instruction is accepted but a change will be made, such as date or remittance not sent
-   * * `ACWP` - AcceptedWithoutPosting. Payment instruction included in the credit transfer is accepted without being posted to the creditor customer’s account
-   * * `RCVD` - Received. Payment initiation has been received by the receiving agent
-   * * `PDNG` - Pending. Payment initiation or individual transaction included in the payment initiation is pending. Further checks and status update will be performed
-   * * `RJCT` - Rejected. Payment initiation or individual transaction included in the payment initiation has been rejected.
-   * * `CANC` - Cancelled. Payment initiation has been cancelled before execution
-   * * `ACFC` - AcceptedFundsChecked. Pre-ceeding check of technical validation and customer profile was successful and an automatic funds check was positive
-   * * `PATC` - PartiallyAcceptedTechnicalCorrect. The payment initiation needs multiple authentications, where some but not yet all have been performed
-   * * `PART` - PartiallyAccepted. A number of transactions have been accepted, whereas another number of transactions have not yet achieved 'accepted' status
-   */
-  payment_status?: PaymentStatusEnum;
-  /**
-   * Payment Type
-   *
-   * * `single-payment` - payment
-   * * `bulk-payment` - bulk-payments
-   * * `periodic-payment` - periodic-payments
-   */
-  payment_type?: PaymentTypeEnum;
-  /**
-   * Redirect URL to your application after payment is done
-   * @format uri
-   * @maxLength 1024
-   */
-  redirect: string | null;
-  /**
-   * Payment Execution date (for periodic payments)
-   * @format date
-   */
-  requested_execution_date?: string;
-  /**
-   * Indicates whether payment should be submitted separately
-   * @default false
-   */
-  submit_payment?: boolean;
-}
-
-export interface ISCT {
-  /** Creditor account */
-  creditor_object: CreditorAccountWrite;
-  /**
-   * Payment end to end identification.
-   * Payment Custom Payment ID
-   * @maxLength 35
-   */
-  custom_payment_id?: string;
-  /**
-   * Payment description
-   * @default "GOCARDLESS"
-   */
-  description?: string;
-  /** Instructed amount */
-  instructed_amount: InstructedAmount;
-  /**
-   * Link to initiate authorization with Institution
-   * @format uri
-   * @default "https://ob.nordigen.com/pis/start/1b1b329f-a654-4ed3-b244-8245de396aa8/{$INSTITUTION_ID}"
-   */
-  link?: string;
-  /** Payment ID */
-  payment_id?: string;
-  /**
-   * Payment end to end identification
-   *
-   * * `INIT` - Initiated. Payment has been initiated.
-   * * `ERRE` - ExecutionError. We experienced error on payment execution.
-   * * `ERRS` - StatusError. We experienced error retrieving payment status. Try again.
-   * * `ACCC` - AcceptedSettlementCompleted. Settlement on the creditor's account has been completed
-   * * `ACCP` - AcceptedCustomerProfile. Preceding check of technical validation was successful. Customer profile check was successful
-   * * `ACSC` - AcceptedSettlementCompleted. Settlement on the debtor’s account has been completed
-   * * `ACSP` - AcceptedSettlementInProcess. All preceding checks such as technical validation and customer profile were successful and therefore the payment initiation has been accepted for execution
-   * * `ACTC` - AcceptedTechnicalValidation. Authentication and syntactical and semantical validation are successful
-   * * `ACWC` - AcceptedWithChange. Instruction is accepted but a change will be made, such as date or remittance not sent
-   * * `ACWP` - AcceptedWithoutPosting. Payment instruction included in the credit transfer is accepted without being posted to the creditor customer’s account
-   * * `RCVD` - Received. Payment initiation has been received by the receiving agent
-   * * `PDNG` - Pending. Payment initiation or individual transaction included in the payment initiation is pending. Further checks and status update will be performed
-   * * `RJCT` - Rejected. Payment initiation or individual transaction included in the payment initiation has been rejected.
-   * * `CANC` - Cancelled. Payment initiation has been cancelled before execution
-   * * `ACFC` - AcceptedFundsChecked. Pre-ceeding check of technical validation and customer profile was successful and an automatic funds check was positive
-   * * `PATC` - PartiallyAcceptedTechnicalCorrect. The payment initiation needs multiple authentications, where some but not yet all have been performed
-   * * `PART` - PartiallyAccepted. A number of transactions have been accepted, whereas another number of transactions have not yet achieved 'accepted' status
-   */
-  payment_status?: PaymentStatusEnum;
-  /**
-   * Payment Type
-   *
-   * * `single-payment` - payment
-   * * `bulk-payment` - bulk-payments
-   * * `periodic-payment` - periodic-payments
-   */
-  payment_type?: PaymentTypeEnum;
-  /**
-   * Redirect URL to your application after payment is done
-   * @format uri
-   * @maxLength 1024
-   */
-  redirect: string | null;
-  /**
-   * Payment Execution date (for periodic payments)
-   * @format date
-   */
-  requested_execution_date?: string;
-  /**
-   * Indicates whether payment should be submitted separately
-   * @default false
-   */
-  submit_payment?: boolean;
-}
-
-/** InstructedAmountSerializer. */
-export interface InstructedAmount {
-  /**
-   * Instructed amount
-   * @format decimal
-   * @pattern ^-?\d{0,10}(?:\.\d{0,2})?$
-   */
-  amount: string;
-  /**
-   * Instructed amount currency
-   * @maxLength 3
-   */
-  currency: string;
-}
-
-/** InstructedAmountSerializer. */
-export interface InstructedAmountRequest {
-  /**
-   * Instructed amount
-   * @format decimal
-   * @pattern ^-?\d{0,10}(?:\.\d{0,2})?$
-   */
-  amount: string;
-  /**
-   * Instructed amount currency
-   * @minLength 1
-   * @maxLength 3
-   */
-  currency: string;
-}
 
 /** Represents an Integration. */
 export interface Integration {
@@ -1536,6 +287,7 @@ export interface Integration {
   countries: string[];
   id: string;
   logo: string;
+  max_access_valid_for_days?: string;
   name: string;
   /** @default "90" */
   transaction_total_days?: string;
@@ -1548,6 +300,7 @@ export interface IntegrationRetrieve {
   id: string;
   identification_codes: any[];
   logo: string;
+  max_access_valid_for_days?: string;
   name: string;
   supported_features: any[];
   supported_payments: Record<string, any>;
@@ -1575,104 +328,11 @@ export interface JWTRefreshRequest {
   refresh: string;
 }
 
-export interface MT {
-  /** Creditor account */
-  creditor_object: CreditorAccountWrite;
-  /**
-   * Payment end to end identification.
-   * Payment Custom Payment ID
-   * @maxLength 35
-   */
-  custom_payment_id?: string;
-  /**
-   * Payment description
-   * @default "GOCARDLESS"
-   */
-  description?: string;
-  /** Instructed amount */
-  instructed_amount: InstructedAmount;
-  /**
-   * Link to initiate authorization with Institution
-   * @format uri
-   * @default "https://ob.nordigen.com/pis/start/1b1b329f-a654-4ed3-b244-8245de396aa8/{$INSTITUTION_ID}"
-   */
-  link?: string;
-  /** Payment ID */
-  payment_id?: string;
-  /**
-   * Payment product
-   *
-   * * `T2P` - target-2-payments
-   * * `SCT` - sepa-credit-transfers
-   * * `ISCT` - instant-sepa-credit-transfer
-   * * `CBCT` - cross-border-credit-transfers
-   * * `BACS` - Back Payment Scheme
-   * * `CHAPS` - CHAPS Payment Scheme
-   * * `FPS` - Faster Payment Scheme
-   * * `SWIFT` - Swift Payment Service
-   * * `BT` - Balance Transfer
-   * * `MT` - Money Transfer
-   * * `DCT` - domestic-credit-transfer
-   * * `IDCT` - instant-domestic-credit-transfer
-   * @default "ISCT"
-   */
-  payment_product?: PaymentProductEnum;
-  /**
-   * Payment end to end identification
-   *
-   * * `INIT` - Initiated. Payment has been initiated.
-   * * `ERRE` - ExecutionError. We experienced error on payment execution.
-   * * `ERRS` - StatusError. We experienced error retrieving payment status. Try again.
-   * * `ACCC` - AcceptedSettlementCompleted. Settlement on the creditor's account has been completed
-   * * `ACCP` - AcceptedCustomerProfile. Preceding check of technical validation was successful. Customer profile check was successful
-   * * `ACSC` - AcceptedSettlementCompleted. Settlement on the debtor’s account has been completed
-   * * `ACSP` - AcceptedSettlementInProcess. All preceding checks such as technical validation and customer profile were successful and therefore the payment initiation has been accepted for execution
-   * * `ACTC` - AcceptedTechnicalValidation. Authentication and syntactical and semantical validation are successful
-   * * `ACWC` - AcceptedWithChange. Instruction is accepted but a change will be made, such as date or remittance not sent
-   * * `ACWP` - AcceptedWithoutPosting. Payment instruction included in the credit transfer is accepted without being posted to the creditor customer’s account
-   * * `RCVD` - Received. Payment initiation has been received by the receiving agent
-   * * `PDNG` - Pending. Payment initiation or individual transaction included in the payment initiation is pending. Further checks and status update will be performed
-   * * `RJCT` - Rejected. Payment initiation or individual transaction included in the payment initiation has been rejected.
-   * * `CANC` - Cancelled. Payment initiation has been cancelled before execution
-   * * `ACFC` - AcceptedFundsChecked. Pre-ceeding check of technical validation and customer profile was successful and an automatic funds check was positive
-   * * `PATC` - PartiallyAcceptedTechnicalCorrect. The payment initiation needs multiple authentications, where some but not yet all have been performed
-   * * `PART` - PartiallyAccepted. A number of transactions have been accepted, whereas another number of transactions have not yet achieved 'accepted' status
-   */
-  payment_status?: PaymentStatusEnum;
-  /**
-   * Payment Type
-   *
-   * * `single-payment` - payment
-   * * `bulk-payment` - bulk-payments
-   * * `periodic-payment` - periodic-payments
-   */
-  payment_type?: PaymentTypeEnum;
-  /**
-   * Redirect URL to your application after payment is done
-   * @format uri
-   * @maxLength 1024
-   */
-  redirect: string | null;
-  /**
-   * Payment Execution date (for periodic payments)
-   * @format date
-   */
-  requested_execution_date?: string;
-  /**
-   * Indicates whether payment should be submitted separately
-   * @default false
-   */
-  submit_payment?: boolean;
-}
-
-export interface MinimumRequiredFields {
-  "single-payment": SinglePayment;
-}
-
 export type ObtainNewAccessRefreshTokenPairData = SpectacularJWTObtain;
 
 export type ObtainNewAccessRefreshTokenPairError = ErrorResponse;
 
+/** OwnerAddressStructuredSchema. */
 export interface OwnerAddressStructuredSchema {
   /** buildingNumber */
   buildingNumber?: string;
@@ -1686,25 +346,9 @@ export interface OwnerAddressStructuredSchema {
   townName?: string;
 }
 
-export interface PaginatedCreditorAccountList {
-  /** @example 123 */
-  count?: number;
-  /**
-   * @format uri
-   * @example "http://api.example.org/accounts/?offset=400&limit=100"
-   */
-  next?: string | null;
-  /**
-   * @format uri
-   * @example "http://api.example.org/accounts/?offset=200&limit=100"
-   */
-  previous?: string | null;
-  results?: CreditorAccount[];
-}
-
 export interface PaginatedEndUserAgreementList {
   /** @example 123 */
-  count?: number;
+  count: number;
   /**
    * @format uri
    * @example "https://bankaccountdata.gocardless.com/api/v2/agreements/enduser/?limit=100&offset=0"
@@ -1715,28 +359,12 @@ export interface PaginatedEndUserAgreementList {
    * @example "https://bankaccountdata.gocardless.com/api/v2/agreements/enduser/?limit=100&offset=0"
    */
   previous?: string | null;
-  results?: EndUserAgreement[];
-}
-
-export interface PaginatedPaymentReadList {
-  /** @example 123 */
-  count?: number;
-  /**
-   * @format uri
-   * @example "http://api.example.org/accounts/?offset=400&limit=100"
-   */
-  next?: string | null;
-  /**
-   * @format uri
-   * @example "http://api.example.org/accounts/?offset=200&limit=100"
-   */
-  previous?: string | null;
-  results?: PaymentRead[];
+  results: EndUserAgreement[];
 }
 
 export interface PaginatedRequisitionList {
   /** @example 123 */
-  count?: number;
+  count: number;
   /**
    * @format uri
    * @example "https://bankaccountdata.gocardless.com/api/v2/requisitions/?limit=100&offset=0"
@@ -1747,461 +375,7 @@ export interface PaginatedRequisitionList {
    * @example "https://bankaccountdata.gocardless.com/api/v2/requisitions/?limit=100&offset=0"
    */
   previous?: string | null;
-  results?: Requisition[];
-}
-
-export interface PaymentDeleted {
-  detail: string;
-  summary: string;
-}
-
-/**
- * * `T2P` - target-2-payments
- * * `SCT` - sepa-credit-transfers
- * * `ISCT` - instant-sepa-credit-transfer
- * * `CBCT` - cross-border-credit-transfers
- * * `BACS` - Back Payment Scheme
- * * `CHAPS` - CHAPS Payment Scheme
- * * `FPS` - Faster Payment Scheme
- * * `SWIFT` - Swift Payment Service
- * * `BT` - Balance Transfer
- * * `MT` - Money Transfer
- * * `DCT` - domestic-credit-transfer
- * * `IDCT` - instant-domestic-credit-transfer
- */
-export enum PaymentProductEnum {
-  T2P = "T2P",
-  SCT = "SCT",
-  ISCT = "ISCT",
-  CBCT = "CBCT",
-  BACS = "BACS",
-  CHAPS = "CHAPS",
-  FPS = "FPS",
-  SWIFT = "SWIFT",
-  BT = "BT",
-  MT = "MT",
-  DCT = "DCT",
-  IDCT = "IDCT",
-}
-
-/** PaymentReadSerializer. */
-export interface PaymentRead {
-  /** Creditor account */
-  creditor_object: CreditorAccountWrite;
-  /**
-   * Payment end to end identification.
-   * Payment Custom Payment ID
-   * @maxLength 35
-   */
-  custom_payment_id?: string;
-  /** Debtor account write serializer. */
-  debtor_account: DebtorAccountWrite;
-  /**
-   * Payment description
-   * @default "GOCARDLESS"
-   */
-  description?: string;
-  /** Instructed amount */
-  instructed_amount: InstructedAmount;
-  /** Payment ID */
-  payment_id?: string;
-  /**
-   * Payment product
-   *
-   * * `T2P` - target-2-payments
-   * * `SCT` - sepa-credit-transfers
-   * * `ISCT` - instant-sepa-credit-transfer
-   * * `CBCT` - cross-border-credit-transfers
-   * * `BACS` - Back Payment Scheme
-   * * `CHAPS` - CHAPS Payment Scheme
-   * * `FPS` - Faster Payment Scheme
-   * * `SWIFT` - Swift Payment Service
-   * * `BT` - Balance Transfer
-   * * `MT` - Money Transfer
-   * * `DCT` - domestic-credit-transfer
-   * * `IDCT` - instant-domestic-credit-transfer
-   * @default "ISCT"
-   */
-  payment_product?: PaymentProductEnum;
-  /**
-   * Payment end to end identification
-   *
-   * * `INIT` - Initiated. Payment has been initiated.
-   * * `ERRE` - ExecutionError. We experienced error on payment execution.
-   * * `ERRS` - StatusError. We experienced error retrieving payment status. Try again.
-   * * `ACCC` - AcceptedSettlementCompleted. Settlement on the creditor's account has been completed
-   * * `ACCP` - AcceptedCustomerProfile. Preceding check of technical validation was successful. Customer profile check was successful
-   * * `ACSC` - AcceptedSettlementCompleted. Settlement on the debtor’s account has been completed
-   * * `ACSP` - AcceptedSettlementInProcess. All preceding checks such as technical validation and customer profile were successful and therefore the payment initiation has been accepted for execution
-   * * `ACTC` - AcceptedTechnicalValidation. Authentication and syntactical and semantical validation are successful
-   * * `ACWC` - AcceptedWithChange. Instruction is accepted but a change will be made, such as date or remittance not sent
-   * * `ACWP` - AcceptedWithoutPosting. Payment instruction included in the credit transfer is accepted without being posted to the creditor customer’s account
-   * * `RCVD` - Received. Payment initiation has been received by the receiving agent
-   * * `PDNG` - Pending. Payment initiation or individual transaction included in the payment initiation is pending. Further checks and status update will be performed
-   * * `RJCT` - Rejected. Payment initiation or individual transaction included in the payment initiation has been rejected.
-   * * `CANC` - Cancelled. Payment initiation has been cancelled before execution
-   * * `ACFC` - AcceptedFundsChecked. Pre-ceeding check of technical validation and customer profile was successful and an automatic funds check was positive
-   * * `PATC` - PartiallyAcceptedTechnicalCorrect. The payment initiation needs multiple authentications, where some but not yet all have been performed
-   * * `PART` - PartiallyAccepted. A number of transactions have been accepted, whereas another number of transactions have not yet achieved 'accepted' status
-   */
-  payment_status?: PaymentStatusEnum;
-  /**
-   * Payment Type
-   *
-   * * `single-payment` - payment
-   * * `bulk-payment` - bulk-payments
-   * * `periodic-payment` - periodic-payments
-   */
-  payment_type?: PaymentTypeEnum;
-  /**
-   * Redirect URL to your application after payment is done
-   * @format uri
-   * @maxLength 1024
-   */
-  redirect: string | null;
-}
-
-/** PaymentReadSerializer. */
-export interface PaymentReadRequest {
-  /** Creditor account */
-  creditor_object: CreditorAccountWriteRequest;
-  /**
-   * Payment end to end identification.
-   * Payment Custom Payment ID
-   * @minLength 1
-   * @maxLength 35
-   */
-  custom_payment_id?: string;
-  /** Debtor account write serializer. */
-  debtor_account: DebtorAccountWriteRequest;
-  /**
-   * Payment description
-   * @minLength 1
-   * @default "GOCARDLESS"
-   */
-  description?: string;
-  /**
-   * Institution ID for Payment
-   * @minLength 1
-   */
-  institution_id: string;
-  /** Instructed amount */
-  instructed_amount: InstructedAmountRequest;
-  /**
-   * Payment product
-   *
-   * * `T2P` - target-2-payments
-   * * `SCT` - sepa-credit-transfers
-   * * `ISCT` - instant-sepa-credit-transfer
-   * * `CBCT` - cross-border-credit-transfers
-   * * `BACS` - Back Payment Scheme
-   * * `CHAPS` - CHAPS Payment Scheme
-   * * `FPS` - Faster Payment Scheme
-   * * `SWIFT` - Swift Payment Service
-   * * `BT` - Balance Transfer
-   * * `MT` - Money Transfer
-   * * `DCT` - domestic-credit-transfer
-   * * `IDCT` - instant-domestic-credit-transfer
-   * @default "ISCT"
-   */
-  payment_product?: PaymentProductEnum;
-  /**
-   * Redirect URL to your application after payment is done
-   * @format uri
-   * @minLength 1
-   * @maxLength 1024
-   */
-  redirect: string | null;
-}
-
-/**
- * * `INIT` - Initiated. Payment has been initiated.
- * * `ERRE` - ExecutionError. We experienced error on payment execution.
- * * `ERRS` - StatusError. We experienced error retrieving payment status. Try again.
- * * `ACCC` - AcceptedSettlementCompleted. Settlement on the creditor's account has been completed
- * * `ACCP` - AcceptedCustomerProfile. Preceding check of technical validation was successful. Customer profile check was successful
- * * `ACSC` - AcceptedSettlementCompleted. Settlement on the debtor’s account has been completed
- * * `ACSP` - AcceptedSettlementInProcess. All preceding checks such as technical validation and customer profile were successful and therefore the payment initiation has been accepted for execution
- * * `ACTC` - AcceptedTechnicalValidation. Authentication and syntactical and semantical validation are successful
- * * `ACWC` - AcceptedWithChange. Instruction is accepted but a change will be made, such as date or remittance not sent
- * * `ACWP` - AcceptedWithoutPosting. Payment instruction included in the credit transfer is accepted without being posted to the creditor customer’s account
- * * `RCVD` - Received. Payment initiation has been received by the receiving agent
- * * `PDNG` - Pending. Payment initiation or individual transaction included in the payment initiation is pending. Further checks and status update will be performed
- * * `RJCT` - Rejected. Payment initiation or individual transaction included in the payment initiation has been rejected.
- * * `CANC` - Cancelled. Payment initiation has been cancelled before execution
- * * `ACFC` - AcceptedFundsChecked. Pre-ceeding check of technical validation and customer profile was successful and an automatic funds check was positive
- * * `PATC` - PartiallyAcceptedTechnicalCorrect. The payment initiation needs multiple authentications, where some but not yet all have been performed
- * * `PART` - PartiallyAccepted. A number of transactions have been accepted, whereas another number of transactions have not yet achieved 'accepted' status
- */
-export enum PaymentStatusEnum {
-  INIT = "INIT",
-  ERRE = "ERRE",
-  ERRS = "ERRS",
-  ACCC = "ACCC",
-  ACCP = "ACCP",
-  ACSC = "ACSC",
-  ACSP = "ACSP",
-  ACTC = "ACTC",
-  ACWC = "ACWC",
-  ACWP = "ACWP",
-  RCVD = "RCVD",
-  PDNG = "PDNG",
-  RJCT = "RJCT",
-  CANC = "CANC",
-  ACFC = "ACFC",
-  PATC = "PATC",
-  PART = "PART",
-}
-
-/**
- * * `single-payment` - payment
- * * `bulk-payment` - bulk-payments
- * * `periodic-payment` - periodic-payments
- */
-export enum PaymentTypeEnum {
-  SinglePayment = "single-payment",
-  BulkPayment = "bulk-payment",
-  PeriodicPayment = "periodic-payment",
-}
-
-/** PaymentWriteSerializer. */
-export interface PaymentWrite {
-  /** Creditor account */
-  creditor_object: CreditorAccountWrite;
-  /**
-   * Payment end to end identification.
-   * Payment Custom Payment ID
-   * @maxLength 35
-   */
-  custom_payment_id?: string;
-  /**
-   * Payment description
-   * @default "GOCARDLESS"
-   */
-  description?: string;
-  /** Instructed amount */
-  instructed_amount: InstructedAmount;
-  /**
-   * Link to initiate authorization with Institution
-   * @format uri
-   * @default "https://ob.nordigen.com/pis/start/1b1b329f-a654-4ed3-b244-8245de396aa8/{$INSTITUTION_ID}"
-   */
-  link?: string;
-  /** Payment ID */
-  payment_id?: string;
-  /**
-   * Payment product
-   *
-   * * `T2P` - target-2-payments
-   * * `SCT` - sepa-credit-transfers
-   * * `ISCT` - instant-sepa-credit-transfer
-   * * `CBCT` - cross-border-credit-transfers
-   * * `BACS` - Back Payment Scheme
-   * * `CHAPS` - CHAPS Payment Scheme
-   * * `FPS` - Faster Payment Scheme
-   * * `SWIFT` - Swift Payment Service
-   * * `BT` - Balance Transfer
-   * * `MT` - Money Transfer
-   * * `DCT` - domestic-credit-transfer
-   * * `IDCT` - instant-domestic-credit-transfer
-   * @default "ISCT"
-   */
-  payment_product?: PaymentProductEnum;
-  /**
-   * Payment end to end identification
-   *
-   * * `INIT` - Initiated. Payment has been initiated.
-   * * `ERRE` - ExecutionError. We experienced error on payment execution.
-   * * `ERRS` - StatusError. We experienced error retrieving payment status. Try again.
-   * * `ACCC` - AcceptedSettlementCompleted. Settlement on the creditor's account has been completed
-   * * `ACCP` - AcceptedCustomerProfile. Preceding check of technical validation was successful. Customer profile check was successful
-   * * `ACSC` - AcceptedSettlementCompleted. Settlement on the debtor’s account has been completed
-   * * `ACSP` - AcceptedSettlementInProcess. All preceding checks such as technical validation and customer profile were successful and therefore the payment initiation has been accepted for execution
-   * * `ACTC` - AcceptedTechnicalValidation. Authentication and syntactical and semantical validation are successful
-   * * `ACWC` - AcceptedWithChange. Instruction is accepted but a change will be made, such as date or remittance not sent
-   * * `ACWP` - AcceptedWithoutPosting. Payment instruction included in the credit transfer is accepted without being posted to the creditor customer’s account
-   * * `RCVD` - Received. Payment initiation has been received by the receiving agent
-   * * `PDNG` - Pending. Payment initiation or individual transaction included in the payment initiation is pending. Further checks and status update will be performed
-   * * `RJCT` - Rejected. Payment initiation or individual transaction included in the payment initiation has been rejected.
-   * * `CANC` - Cancelled. Payment initiation has been cancelled before execution
-   * * `ACFC` - AcceptedFundsChecked. Pre-ceeding check of technical validation and customer profile was successful and an automatic funds check was positive
-   * * `PATC` - PartiallyAcceptedTechnicalCorrect. The payment initiation needs multiple authentications, where some but not yet all have been performed
-   * * `PART` - PartiallyAccepted. A number of transactions have been accepted, whereas another number of transactions have not yet achieved 'accepted' status
-   */
-  payment_status?: PaymentStatusEnum;
-  /**
-   * Payment Type
-   *
-   * * `single-payment` - payment
-   * * `bulk-payment` - bulk-payments
-   * * `periodic-payment` - periodic-payments
-   */
-  payment_type?: PaymentTypeEnum;
-  /** Periodic Payment Serializer. */
-  periodic_payment?: PeriodicPayment;
-  /**
-   * Redirect URL to your application after payment is done
-   * @format uri
-   * @maxLength 1024
-   */
-  redirect: string | null;
-  /**
-   * Payment Execution date (for periodic payments)
-   * @format date
-   */
-  requested_execution_date?: string;
-  /**
-   * Indicates whether payment should be submitted separately
-   * @default false
-   */
-  submit_payment?: boolean;
-}
-
-/** PaymentWriteSerializer. */
-export interface PaymentWriteRequest {
-  /** Creditor account */
-  creditor_object: CreditorAccountWriteRequest;
-  /**
-   * Payment end to end identification.
-   * Payment Custom Payment ID
-   * @minLength 1
-   * @maxLength 35
-   */
-  custom_payment_id?: string;
-  /** Debtor account */
-  debtor_account?: DebtorAccountWriteRequest;
-  /**
-   * Payment description
-   * @minLength 1
-   * @default "GOCARDLESS"
-   */
-  description?: string;
-  /**
-   * Institution ID for Payment
-   * @minLength 1
-   */
-  institution_id: string;
-  /** Instructed amount */
-  instructed_amount: InstructedAmountRequest;
-  /**
-   * Payment product
-   *
-   * * `T2P` - target-2-payments
-   * * `SCT` - sepa-credit-transfers
-   * * `ISCT` - instant-sepa-credit-transfer
-   * * `CBCT` - cross-border-credit-transfers
-   * * `BACS` - Back Payment Scheme
-   * * `CHAPS` - CHAPS Payment Scheme
-   * * `FPS` - Faster Payment Scheme
-   * * `SWIFT` - Swift Payment Service
-   * * `BT` - Balance Transfer
-   * * `MT` - Money Transfer
-   * * `DCT` - domestic-credit-transfer
-   * * `IDCT` - instant-domestic-credit-transfer
-   * @default "ISCT"
-   */
-  payment_product?: PaymentProductEnum;
-  /** Periodic Payment Serializer. */
-  periodic_payment?: PeriodicPaymentRequest;
-  /**
-   * Redirect URL to your application after payment is done
-   * @format uri
-   * @minLength 1
-   * @maxLength 1024
-   */
-  redirect: string | null;
-  /**
-   * Payment Execution date (for periodic payments)
-   * @format date
-   */
-  requested_execution_date?: string;
-  /**
-   * Indicates whether payment should be submitted separately
-   * @default false
-   */
-  submit_payment?: boolean;
-}
-
-/** Periodic Payment Serializer. */
-export interface PeriodicPayment {
-  day_of_execution?: string;
-  /** @format date */
-  end_date?: string;
-  /**
-   * Behavior when periodic payment dates fall on holiday.
-   *
-   * * `following` - following
-   * * `preceding` - preceding
-   * @default "following"
-   */
-  execution_rule?: ExecutionRuleEnum;
-  /** @default "Monthly" */
-  frequency?: FrequencyEnum;
-  /** @format date */
-  start_date: string;
-}
-
-/** Periodic Payment Serializer. */
-export interface PeriodicPaymentRequest {
-  /** @minLength 1 */
-  day_of_execution?: string;
-  /** @format date */
-  end_date?: string;
-  /**
-   * Behavior when periodic payment dates fall on holiday.
-   *
-   * * `following` - following
-   * * `preceding` - preceding
-   * @default "following"
-   */
-  execution_rule?: ExecutionRuleEnum;
-  /** @default "Monthly" */
-  frequency?: FrequencyEnum;
-  /** @format date */
-  start_date: string;
-}
-
-/** Filter country. */
-export interface PremiumAccountQuery {
-  /**
-   * * `AT` - Austria
-   * * `BE` - Belgium
-   * * `BG` - Bulgaria
-   * * `HR` - Croatia
-   * * `CY` - Cyprus
-   * * `CZ` - Czechia
-   * * `DK` - Denmark
-   * * `EE` - Estonia
-   * * `FI` - Finland
-   * * `FR` - France
-   * * `DE` - Germany
-   * * `GR` - Greece
-   * * `HU` - Hungary
-   * * `IS` - Iceland
-   * * `IE` - Ireland
-   * * `IT` - Italy
-   * * `LV` - Latvia
-   * * `LI` - Liechtenstein
-   * * `LT` - Lithuania
-   * * `LU` - Luxembourg
-   * * `MT` - Malta
-   * * `NL` - Netherlands
-   * * `NO` - Norway
-   * * `PL` - Poland
-   * * `PT` - Portugal
-   * * `RO` - Romania
-   * * `SK` - Slovakia
-   * * `SI` - Slovenia
-   * * `ES` - Spain
-   * * `SE` - Sweden
-   * * `GB` - United Kingdom
-   * * `US` - United States of America
-   */
-  country?: CountryEnum;
-  /** @format date */
-  date_from?: string;
-  /** @format date */
-  date_to?: string;
+  results: Requisition[];
 }
 
 /** RequisitionSerializer. */
@@ -2231,7 +405,7 @@ export interface Requisition {
   /**
    * link to initiate authorization with Institution
    * @format uri
-   * @default "https://ob.nordigen.com/psd2/start/3fa85f64-5717-4562-b3fc-2c963f66afa6/{$INSTITUTION_ID}"
+   * @default "https://ob.gocardless.com/psd2/start/3fa85f64-5717-4562-b3fc-2c963f66afa6/SANDBOXFINANCE_SFIN0000"
    */
   link?: string;
   /**
@@ -2259,7 +433,7 @@ export interface Requisition {
    * Requisition status
    * status of this requisition
    */
-  status?: Status1C5Enum;
+  status?: StatusEnum;
   /**
    * A two-letter country code (ISO 639-1)
    * @maxLength 5
@@ -2319,11 +493,11 @@ export interface RequisitionRequest {
   user_language?: string;
 }
 
-export type RetrieveAccountBalancesData = AccountBalanceSchema;
+export type RetrieveAccountBalancesData = AccountBalance;
 
 export type RetrieveAccountBalancesError = ErrorResponse;
 
-export type RetrieveAccountDetailsData = AccountDetailSchema;
+export type RetrieveAccountDetailsData = AccountDetail;
 
 export type RetrieveAccountDetailsError = ErrorResponse;
 
@@ -2331,7 +505,7 @@ export type RetrieveAccountMetadataData = Account;
 
 export type RetrieveAccountMetadataError = ErrorResponse;
 
-export type RetrieveAccountTransactionsData = BankTransactionStatusSchema;
+export type RetrieveAccountTransactionsData = AccountTransactions;
 
 export type RetrieveAccountTransactionsError = ErrorResponse;
 
@@ -2340,28 +514,6 @@ export interface RetrieveAccountTransactionsParams {
   date_from?: string;
   /** @format date */
   date_to?: string;
-  /**
-   * A UUID string identifying this account.
-   * @format uuid
-   */
-  id: string;
-}
-
-export type RetrieveAccountTransactionsPremiumData = BankTransactionStatusSchema;
-
-export type RetrieveAccountTransactionsPremiumError = ErrorResponse;
-
-export interface RetrieveAccountTransactionsPremiumParams {
-  /** ISO 3166 two-character country code */
-  country?: string;
-  /** @format date */
-  date_from?: string;
-  /** @format date */
-  date_to?: string;
-  /**
-   * A UUID string identifying this account.
-   * @format uuid
-   */
   id: string;
 }
 
@@ -2377,9 +529,9 @@ export interface RetrieveAllEuAsForAnEndUserParams {
    */
   limit?: number;
   /**
-   * The initial index from which to return the results.
-   * @min 1
-   * @default 1
+   * The initial zero-based index from which to return the results.
+   * @min 0
+   * @default 0
    */
   offset?: number;
 }
@@ -2396,9 +548,9 @@ export interface RetrieveAllRequisitionsParams {
    */
   limit?: number;
   /**
-   * The initial index from which to return the results.
-   * @min 1
-   * @default 1
+   * The initial zero-based index from which to return the results.
+   * @min 0
+   * @default 0
    */
   offset?: number;
 }
@@ -2443,201 +595,6 @@ export type RetrieveEuaByIdError = ErrorResponse;
 export type RetrieveInstitutionData = IntegrationRetrieve;
 
 export type RetrieveInstitutionError = ErrorResponse;
-
-export interface SCT {
-  /** Creditor account */
-  creditor_object: CreditorAccountWrite;
-  /**
-   * Payment end to end identification.
-   * Payment Custom Payment ID
-   * @maxLength 35
-   */
-  custom_payment_id?: string;
-  /**
-   * Payment description
-   * @default "GOCARDLESS"
-   */
-  description?: string;
-  /** Instructed amount */
-  instructed_amount: InstructedAmount;
-  /**
-   * Link to initiate authorization with Institution
-   * @format uri
-   * @default "https://ob.nordigen.com/pis/start/1b1b329f-a654-4ed3-b244-8245de396aa8/{$INSTITUTION_ID}"
-   */
-  link?: string;
-  /** Payment ID */
-  payment_id?: string;
-  /**
-   * Payment product
-   *
-   * * `T2P` - target-2-payments
-   * * `SCT` - sepa-credit-transfers
-   * * `ISCT` - instant-sepa-credit-transfer
-   * * `CBCT` - cross-border-credit-transfers
-   * * `BACS` - Back Payment Scheme
-   * * `CHAPS` - CHAPS Payment Scheme
-   * * `FPS` - Faster Payment Scheme
-   * * `SWIFT` - Swift Payment Service
-   * * `BT` - Balance Transfer
-   * * `MT` - Money Transfer
-   * * `DCT` - domestic-credit-transfer
-   * * `IDCT` - instant-domestic-credit-transfer
-   * @default "ISCT"
-   */
-  payment_product?: PaymentProductEnum;
-  /**
-   * Payment end to end identification
-   *
-   * * `INIT` - Initiated. Payment has been initiated.
-   * * `ERRE` - ExecutionError. We experienced error on payment execution.
-   * * `ERRS` - StatusError. We experienced error retrieving payment status. Try again.
-   * * `ACCC` - AcceptedSettlementCompleted. Settlement on the creditor's account has been completed
-   * * `ACCP` - AcceptedCustomerProfile. Preceding check of technical validation was successful. Customer profile check was successful
-   * * `ACSC` - AcceptedSettlementCompleted. Settlement on the debtor’s account has been completed
-   * * `ACSP` - AcceptedSettlementInProcess. All preceding checks such as technical validation and customer profile were successful and therefore the payment initiation has been accepted for execution
-   * * `ACTC` - AcceptedTechnicalValidation. Authentication and syntactical and semantical validation are successful
-   * * `ACWC` - AcceptedWithChange. Instruction is accepted but a change will be made, such as date or remittance not sent
-   * * `ACWP` - AcceptedWithoutPosting. Payment instruction included in the credit transfer is accepted without being posted to the creditor customer’s account
-   * * `RCVD` - Received. Payment initiation has been received by the receiving agent
-   * * `PDNG` - Pending. Payment initiation or individual transaction included in the payment initiation is pending. Further checks and status update will be performed
-   * * `RJCT` - Rejected. Payment initiation or individual transaction included in the payment initiation has been rejected.
-   * * `CANC` - Cancelled. Payment initiation has been cancelled before execution
-   * * `ACFC` - AcceptedFundsChecked. Pre-ceeding check of technical validation and customer profile was successful and an automatic funds check was positive
-   * * `PATC` - PartiallyAcceptedTechnicalCorrect. The payment initiation needs multiple authentications, where some but not yet all have been performed
-   * * `PART` - PartiallyAccepted. A number of transactions have been accepted, whereas another number of transactions have not yet achieved 'accepted' status
-   */
-  payment_status?: PaymentStatusEnum;
-  /**
-   * Payment Type
-   *
-   * * `single-payment` - payment
-   * * `bulk-payment` - bulk-payments
-   * * `periodic-payment` - periodic-payments
-   */
-  payment_type?: PaymentTypeEnum;
-  /**
-   * Redirect URL to your application after payment is done
-   * @format uri
-   * @maxLength 1024
-   */
-  redirect: string | null;
-  /**
-   * Payment Execution date (for periodic payments)
-   * @format date
-   */
-  requested_execution_date?: string;
-  /**
-   * Indicates whether payment should be submitted separately
-   * @default false
-   */
-  submit_payment?: boolean;
-}
-
-export interface SWIFT {
-  /** Creditor account */
-  creditor_object: CreditorAccountWrite;
-  /**
-   * Payment end to end identification.
-   * Payment Custom Payment ID
-   * @maxLength 35
-   */
-  custom_payment_id?: string;
-  /**
-   * Payment description
-   * @default "GOCARDLESS"
-   */
-  description?: string;
-  /** Instructed amount */
-  instructed_amount: InstructedAmount;
-  /**
-   * Link to initiate authorization with Institution
-   * @format uri
-   * @default "https://ob.nordigen.com/pis/start/1b1b329f-a654-4ed3-b244-8245de396aa8/{$INSTITUTION_ID}"
-   */
-  link?: string;
-  /** Payment ID */
-  payment_id?: string;
-  /**
-   * Payment product
-   *
-   * * `T2P` - target-2-payments
-   * * `SCT` - sepa-credit-transfers
-   * * `ISCT` - instant-sepa-credit-transfer
-   * * `CBCT` - cross-border-credit-transfers
-   * * `BACS` - Back Payment Scheme
-   * * `CHAPS` - CHAPS Payment Scheme
-   * * `FPS` - Faster Payment Scheme
-   * * `SWIFT` - Swift Payment Service
-   * * `BT` - Balance Transfer
-   * * `MT` - Money Transfer
-   * * `DCT` - domestic-credit-transfer
-   * * `IDCT` - instant-domestic-credit-transfer
-   * @default "ISCT"
-   */
-  payment_product?: PaymentProductEnum;
-  /**
-   * Payment end to end identification
-   *
-   * * `INIT` - Initiated. Payment has been initiated.
-   * * `ERRE` - ExecutionError. We experienced error on payment execution.
-   * * `ERRS` - StatusError. We experienced error retrieving payment status. Try again.
-   * * `ACCC` - AcceptedSettlementCompleted. Settlement on the creditor's account has been completed
-   * * `ACCP` - AcceptedCustomerProfile. Preceding check of technical validation was successful. Customer profile check was successful
-   * * `ACSC` - AcceptedSettlementCompleted. Settlement on the debtor’s account has been completed
-   * * `ACSP` - AcceptedSettlementInProcess. All preceding checks such as technical validation and customer profile were successful and therefore the payment initiation has been accepted for execution
-   * * `ACTC` - AcceptedTechnicalValidation. Authentication and syntactical and semantical validation are successful
-   * * `ACWC` - AcceptedWithChange. Instruction is accepted but a change will be made, such as date or remittance not sent
-   * * `ACWP` - AcceptedWithoutPosting. Payment instruction included in the credit transfer is accepted without being posted to the creditor customer’s account
-   * * `RCVD` - Received. Payment initiation has been received by the receiving agent
-   * * `PDNG` - Pending. Payment initiation or individual transaction included in the payment initiation is pending. Further checks and status update will be performed
-   * * `RJCT` - Rejected. Payment initiation or individual transaction included in the payment initiation has been rejected.
-   * * `CANC` - Cancelled. Payment initiation has been cancelled before execution
-   * * `ACFC` - AcceptedFundsChecked. Pre-ceeding check of technical validation and customer profile was successful and an automatic funds check was positive
-   * * `PATC` - PartiallyAcceptedTechnicalCorrect. The payment initiation needs multiple authentications, where some but not yet all have been performed
-   * * `PART` - PartiallyAccepted. A number of transactions have been accepted, whereas another number of transactions have not yet achieved 'accepted' status
-   */
-  payment_status?: PaymentStatusEnum;
-  /**
-   * Payment Type
-   *
-   * * `single-payment` - payment
-   * * `bulk-payment` - bulk-payments
-   * * `periodic-payment` - periodic-payments
-   */
-  payment_type?: PaymentTypeEnum;
-  /**
-   * Redirect URL to your application after payment is done
-   * @format uri
-   * @maxLength 1024
-   */
-  redirect: string | null;
-  /**
-   * Payment Execution date (for periodic payments)
-   * @format date
-   */
-  requested_execution_date?: string;
-  /**
-   * Indicates whether payment should be submitted separately
-   * @default false
-   */
-  submit_payment?: boolean;
-}
-
-export interface SinglePayment {
-  T2P: T2P;
-  BACS: BACS;
-  BT: BT;
-  CBCT: CBCT;
-  CHAPS: CHAPS;
-  DCT: DCT;
-  FPS: FPS;
-  IDCT: IDCT;
-  ISCT: ISCT;
-  MT: MT;
-  SCT: SCT;
-  SWIFT: SWIFT;
-}
 
 /** Obtain new JWT pair. */
 export interface SpectacularJWTObtain {
@@ -2698,7 +655,7 @@ export interface SpectacularRequisition {
   /**
    * link to initiate authorization with Institution
    * @format uri
-   * @default "https://ob.nordigen.com/psd2/start/3fa85f64-5717-4562-b3fc-2c963f66afa6/{$INSTITUTION_ID}"
+   * @default "https://ob.gocardless.com/psd2/start/3fa85f64-5717-4562-b3fc-2c963f66afa6/SANDBOXFINANCE_SFIN0000"
    */
   link?: string;
   /**
@@ -2726,7 +683,7 @@ export interface SpectacularRequisition {
    * Requisition status
    * status of this requisition
    */
-  status?: Status1C5Enum;
+  status?: StatusEnum;
   /**
    * A two-letter country code (ISO 639-1)
    * @maxLength 5
@@ -2734,7 +691,7 @@ export interface SpectacularRequisition {
   user_language?: string;
 }
 
-export enum Status1C5Enum {
+export enum StatusEnum {
   CR = "CR",
   ID = "ID",
   LN = "LN",
@@ -2748,96 +705,12 @@ export enum Status1C5Enum {
   SA = "SA",
 }
 
-export interface T2P {
-  /** Creditor account */
-  creditor_object: CreditorAccountWrite;
-  /**
-   * Payment end to end identification.
-   * Payment Custom Payment ID
-   * @maxLength 35
-   */
-  custom_payment_id?: string;
-  /**
-   * Payment description
-   * @default "GOCARDLESS"
-   */
-  description?: string;
-  /** Instructed amount */
-  instructed_amount: InstructedAmount;
-  /**
-   * Link to initiate authorization with Institution
-   * @format uri
-   * @default "https://ob.nordigen.com/pis/start/1b1b329f-a654-4ed3-b244-8245de396aa8/{$INSTITUTION_ID}"
-   */
-  link?: string;
-  /** Payment ID */
-  payment_id?: string;
-  /**
-   * Payment product
-   *
-   * * `T2P` - target-2-payments
-   * * `SCT` - sepa-credit-transfers
-   * * `ISCT` - instant-sepa-credit-transfer
-   * * `CBCT` - cross-border-credit-transfers
-   * * `BACS` - Back Payment Scheme
-   * * `CHAPS` - CHAPS Payment Scheme
-   * * `FPS` - Faster Payment Scheme
-   * * `SWIFT` - Swift Payment Service
-   * * `BT` - Balance Transfer
-   * * `MT` - Money Transfer
-   * * `DCT` - domestic-credit-transfer
-   * * `IDCT` - instant-domestic-credit-transfer
-   * @default "ISCT"
-   */
-  payment_product?: PaymentProductEnum;
-  /**
-   * Payment end to end identification
-   *
-   * * `INIT` - Initiated. Payment has been initiated.
-   * * `ERRE` - ExecutionError. We experienced error on payment execution.
-   * * `ERRS` - StatusError. We experienced error retrieving payment status. Try again.
-   * * `ACCC` - AcceptedSettlementCompleted. Settlement on the creditor's account has been completed
-   * * `ACCP` - AcceptedCustomerProfile. Preceding check of technical validation was successful. Customer profile check was successful
-   * * `ACSC` - AcceptedSettlementCompleted. Settlement on the debtor’s account has been completed
-   * * `ACSP` - AcceptedSettlementInProcess. All preceding checks such as technical validation and customer profile were successful and therefore the payment initiation has been accepted for execution
-   * * `ACTC` - AcceptedTechnicalValidation. Authentication and syntactical and semantical validation are successful
-   * * `ACWC` - AcceptedWithChange. Instruction is accepted but a change will be made, such as date or remittance not sent
-   * * `ACWP` - AcceptedWithoutPosting. Payment instruction included in the credit transfer is accepted without being posted to the creditor customer’s account
-   * * `RCVD` - Received. Payment initiation has been received by the receiving agent
-   * * `PDNG` - Pending. Payment initiation or individual transaction included in the payment initiation is pending. Further checks and status update will be performed
-   * * `RJCT` - Rejected. Payment initiation or individual transaction included in the payment initiation has been rejected.
-   * * `CANC` - Cancelled. Payment initiation has been cancelled before execution
-   * * `ACFC` - AcceptedFundsChecked. Pre-ceeding check of technical validation and customer profile was successful and an automatic funds check was positive
-   * * `PATC` - PartiallyAcceptedTechnicalCorrect. The payment initiation needs multiple authentications, where some but not yet all have been performed
-   * * `PART` - PartiallyAccepted. A number of transactions have been accepted, whereas another number of transactions have not yet achieved 'accepted' status
-   */
-  payment_status?: PaymentStatusEnum;
-  /**
-   * Payment Type
-   *
-   * * `single-payment` - payment
-   * * `bulk-payment` - bulk-payments
-   * * `periodic-payment` - periodic-payments
-   */
-  payment_type?: PaymentTypeEnum;
-  /**
-   * Redirect URL to your application after payment is done
-   * @format uri
-   * @maxLength 1024
-   */
-  redirect: string | null;
-  /**
-   * Payment Execution date (for periodic payments)
-   * @format date
-   */
-  requested_execution_date?: string;
-  /**
-   * Indicates whether payment should be submitted separately
-   * @default false
-   */
-  submit_payment?: boolean;
+export interface SuccessfulDeleteResponse {
+  detail: string;
+  summary: string;
 }
 
+/** TransactionAmountSchema. */
 export interface TransactionAmountSchema {
   /** amount */
   amount: string;
@@ -2845,6 +718,7 @@ export interface TransactionAmountSchema {
   currency: string;
 }
 
+/** TransactionSchema. */
 export interface TransactionSchema {
   /** additionalInformation */
   additionalInformation?: string;
@@ -2899,15 +773,4 @@ export interface TransactionSchema {
   valueDate?: string;
   /** valueDateTime */
   valueDateTime?: string;
-}
-
-/**
- * * `IBAN` - IBAN
- * * `SCAN` - SortCodeAccountNumber
- * * `BBAN` - BBAN
- */
-export enum TypeEnum {
-  IBAN = "IBAN",
-  SCAN = "SCAN",
-  BBAN = "BBAN",
 }
