+++ src/index.ts
@@ -1,18 +1,7 @@
 export * from "./NordigenClient.js";
 export * from "./HttpMethod.js";
 export * from "./countries.js";
-export { 
-  AccountStatusEnum, 
-  AddressCountryEnum, 
-  CountryEnum, 
-  ExecutionRuleEnum, 
-  FrequencyEnum,
-  PaymentProductEnum,
-  PaymentStatusEnum,
-  PaymentTypeEnum,
-  Status1C5Enum,
-  TypeEnum,
-} from "./generated.js";
+export * from "./types/enums.js";
 export type * as Api from "./generated.js";
 export type * from "./types/index.js";
 export { NordigenClient as default } from "./NordigenClient.js"
